ONT-2034/2039 # Settings/Signal.

+++ src/main/java/com/specure/core/response/settings/MapServerSettingsResponse.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Data
 public class MapServerSettingsResponse {
     private String host;
-    private int port;
+    private Long port;
     private boolean ssl;
 }
